% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFcor.R
\name{cor_BF10}
\alias{cor_BF10}
\title{Bayes factor for Pearson correlation with informed prior}
\usage{
cor_BF10(N, R, alpha_prior, beta_prior)
}
\arguments{
\item{N}{Sample size}

\item{R}{Pearson correlation coefficient in the sample}

\item{alpha_prior}{Alpha parameter in the beta prior on r}

\item{beta_prior}{Beta parameter in the beta prior on r}
}
\description{
Bayes factor for Pearson correlation with informed prior
}
