% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BFcor_Ly.R
\name{cor_BF10_Ly}
\alias{cor_BF10_Ly}
\title{Bayes factor for Pearson correlation with informed prior}
\usage{
cor_BF10_Ly(N, R, alpha_prior, beta_prior, Gamma = 0, Delta = 0)
}
\arguments{
\item{N}{Sample size}

\item{R}{Pearson correlation coefficient in the sample}

\item{alpha_prior}{Alpha parameter in the beta prior on r}

\item{beta_prior}{Beta parameter in the beta prior on r}

\item{Gamma}{Gamma parameter (see Ly et al. 2018)}

\item{Delta}{Delta parameter (see Ly et al. 2018)}
}
\description{
Bayes factor for Pearson correlation with informed prior
}
