% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFcor_Ly.R
\name{cor_posterior}
\alias{cor_posterior}
\title{Standardized posterior for Pearson correlation with informed prior}
\usage{
cor_posterior(Rho, N, R, alpha_prior, beta_prior)
}
\arguments{
\item{Rho}{Correlation parameter}

\item{N}{Sample size}

\item{R}{Pearson correlation coefficient in the sample}

\item{alpha_prior}{Alpha parameter in the beta prior on r}

\item{beta_prior}{Beta parameter in the beta prior on r}
}
\description{
Standardized posterior for Pearson correlation with informed prior
}
