% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFcor_Ly.R
\name{cor_posterior_Ly}
\alias{cor_posterior_Ly}
\title{Standardized posterior for Pearson correlation with informed prior}
\usage{
cor_posterior_Ly(Rho, N, R, alpha_prior, beta_prior, Gamma, Delta)
}
\arguments{
\item{Rho}{Correlation parameter}

\item{N}{Sample size}

\item{R}{Pearson correlation coefficient in the sample}

\item{alpha_prior}{Alpha parameter in the beta prior on r}

\item{beta_prior}{Beta parameter in the beta prior on r}

\item{Gamma}{Gamma parameter (see Ly et al. 2018)}

\item{Delta}{Delta parameter (see Ly et al. 2018)}
}
\description{
Standardized posterior for Pearson correlation with informed prior
}
