% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dtss}
\alias{dtss}
\title{t-Distribution defined in terms of mean, scale, and df}
\usage{
dtss(x, mu, r, kappa, log = FALSE)
}
\arguments{
\item{x}{Quantile}

\item{mu}{Mean}

\item{r}{Scale parameter}

\item{kappa}{Degrees of freedom}

\item{log}{logical; if TRUE, densities are given as log(p)}
}
\description{
t-Distribution defined in terms of mean, scale, and df
}
