% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeKLcor.R, R/computeKLttest.R
\name{KLBosco_complete}
\alias{KLBosco_complete}
\title{Function to make a random draw from the BoscoEtAl dataset, analyze the correlations, and output the result matrix}
\usage{
KLBosco_complete(BoscoEtAl, ExpertsPriors_cor, seed)

KLBosco_complete(BoscoEtAl, ExpertsPriors_cor, seed)
}
\arguments{
\item{BoscoEtAl}{The Bosco et al. dataset (as constructed in /data-raw)}

\item{ExpertsPriors_cor}{The ExpertsPriors_cor dataset (as constructed in /data-raw)}

\item{seed}{Random seed for random draw of studies from the Bosco et al. dataset}
}
\description{
Function to make a random draw from the BoscoEtAl dataset, analyze the correlations, and output the result matrix

Function to make a random draw from the BoscoEtAl dataset, analyze the correlations, and output the result matrix
}
