% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_KL_Divergence.R
\name{KLdiv_cor}
\alias{KLdiv_cor}
\title{Kullback-Leibler Divergence for the posteriors in the correlation test}
\usage{
KLdiv_cor(N, R, alpha_prior, beta_prior)
}
\arguments{
\item{N}{Sample size}

\item{R}{Correlation in the sample}

\item{alpha_prior}{Alpha parameter of the prior}

\item{beta_prior}{Beta parameter of the prior}
}
\description{
Kullback-Leibler Divergence for the posteriors in the correlation test
}
