% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL_Divergence.R
\name{KLdiv_tTest_t}
\alias{KLdiv_tTest_t}
\title{Kullback-Leibler Divergence for the posteriors in the t-test with t-prior}
\usage{
KLdiv_tTest_t(
  t,
  n1,
  n2,
  independentSamples,
  direction,
  prior.location,
  prior.scale,
  prior.df
)
}
\arguments{
\item{t}{Sample t-value}

\item{n1}{Sample size group1}

\item{n2}{Sample size group2}

\item{independentSamples}{Is it an independent samples t-test?}

\item{direction}{What is the direction of the t-test? greater / less / two.sided}

\item{prior.location}{Location parameter in the prior distribution}

\item{prior.scale}{Scale parameter in the prior distribution}

\item{prior.df}{Degrees of freedom of the prior distribution}
}
\description{
Kullback-Leibler Divergence for the posteriors in the t-test with t-prior
}
