% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BF_all.R
\name{plotallBFBosco}
\alias{plotallBFBosco}
\title{Bayes factor correspondence plot for Bosco et al data}
\usage{
plotallBFBosco(
  BFBosco,
  elicit.stage = "Shiny",
  BFlog = FALSE,
  x1 = 100,
  y1 = 100,
  x0 = 0,
  y0 = 0
)
}
\arguments{
\item{BFBosco}{Dataset containing Bayes factors for Bosco data}

\item{elicit.stage}{"Shiny" or "MATCH" for elicited priors from Shiny app or MATCH}

\item{BFlog}{TRUE/FALSE: Should log BFs be displayed?}

\item{x1}{upper limit x axis}

\item{y1}{upper limit y axis}

\item{x0}{lower limit x axis}

\item{y0}{lower limit y axis}
}
\description{
Bayes factor correspondence plot for Bosco et al data
}
