% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BF_all.R
\name{plotallBFWetzels}
\alias{plotallBFWetzels}
\title{Bayes factor correspondence plot for Wetzels et al data}
\usage{
plotallBFWetzels(
  BFWetzels,
  alternative = "greater",
  elicit.stage = "Shiny",
  disttype = "t",
  BFlog = FALSE,
  x1 = 100,
  y1 = 100,
  x0 = 0,
  y0 = 0
)
}
\arguments{
\item{BFWetzels}{Dataset containing Bayes factors for Wetzels data}

\item{alternative}{one out of "two.sided", "greater", or "less"}

\item{elicit.stage}{"Shiny" or "MATCH" for elicited priors from Shiny app or MATCH}

\item{disttype}{Which prior distribution family? "norm" for normal distribution, "t" for t-distribution}

\item{BFlog}{TRUE/FALSE: Should log BFs be displayed?}

\item{x1}{upper limit x axis}

\item{y1}{upper limit y axis}

\item{x0}{lower limit x axis}

\item{y0}{lower limit y axis}
}
\description{
Bayes factor correspondence plot for Wetzels et al data
}
