% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BF.R
\name{plotBFBoscoES}
\alias{plotBFBoscoES}
\title{Plot Bayes factors for increasing sample effect sizes in the Bosco et al data}
\usage{
plotBFBoscoES(
  BoscoEtAl,
  BFBosco,
  ExpertsPriors_cor,
  elicit.stage,
  percent = 0.3
)
}
\arguments{
\item{BoscoEtAl}{Bosco et al. dataset}

\item{BFBosco}{Dataset containing Bayes factors for Bosco data}

\item{ExpertsPriors_cor}{Dataset ExpertsPriors_tTest containing prior distribution data for the t-test}

\item{elicit.stage}{"Shiny" or "MATCH" for elicited priors from Shiny app or MATCH}

\item{percent}{Determines xlim bounds. Credible interval of the prior distributions.}
}
\description{
Plot Bayes factors for increasing sample effect sizes in the Bosco et al data
}
