% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BF.R
\name{plotBFBoscoH0}
\alias{plotBFBoscoH0}
\title{Plot BF for the Bosco et al dataset where all BF point towards H0}
\usage{
plotBFBoscoH0(BFBosco)
}
\arguments{
\item{BFBosco}{BFBosco dataset containing the BF of the Bosco et al study}
}
\description{
Plot BF for the Bosco et al dataset where all BF point towards H0
}
