% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BF.R
\name{plotBFBoscoH1}
\alias{plotBFBoscoH1}
\title{Plot BF for the Bosco et al dataset where all BF point towards H1}
\usage{
plotBFBoscoH1(BFBosco)
}
\arguments{
\item{BFBosco}{BFBosco dataset containing the BF of the Bosco et al study}
}
\description{
Plot BF for the Bosco et al dataset where all BF point towards H1
}
