% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BF.R
\name{plotBFBoscoMixed}
\alias{plotBFBoscoMixed}
\title{Plot BF for the Bosco et al dataset where BF point to H0 or H1 depending on the prior}
\usage{
plotBFBoscoMixed(BFBosco)
}
\arguments{
\item{BFBosco}{BFBosco dataset containing the BF of the Bosco et al study}
}
\description{
Plot BF for the Bosco et al dataset where BF point to H0 or H1 depending on the prior
}
