% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BFforRs.R
\name{plotBFBoscoOneR}
\alias{plotBFBoscoOneR}
\title{Plot Bayes factors for increasing sample effect sizes in the Bosco et al data}
\usage{
plotBFBoscoOneR(
  BoscoEtAl,
  BFBosco,
  ExpertsPriors_cor,
  elicit.stage,
  whichplots = c(1, 3, 7, 8)
)
}
\arguments{
\item{BoscoEtAl}{Bosco et al. dataset}

\item{BFBosco}{Dataset containing Bayes factors for Bosco data}

\item{ExpertsPriors_cor}{Dataset ExpertsPriors_tTest containing prior distribution data for the t-test}

\item{elicit.stage}{"Shiny" or "MATCH" for elicited priors from Shiny app or MATCH}

\item{whichplots}{For which experts should plots be created? (1=negative ES, 2:7=experts, 8=large positive ES)}
}
\description{
Plot Bayes factors for increasing sample effect sizes in the Bosco et al data
}
