% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BFAgreement.R
\name{plotBFDirectionBosco}
\alias{plotBFDirectionBosco}
\title{Function to plot the agreement in direction between Bayes factors for the Bosco data}
\usage{
plotBFDirectionBosco(BFBosco, elicit.stage)
}
\arguments{
\item{BFBosco}{Dataset containing Bayes factors for Bosco data}

\item{elicit.stage}{Either "MATCH" or "Shiny"}
}
\description{
Function to plot the agreement in direction between Bayes factors for the Bosco data
}
