% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BFAgreement.R
\name{plotBFDirectionWetzels}
\alias{plotBFDirectionWetzels}
\title{Function to plot the agreement in direction between Bayes factors for the Wetzels data}
\usage{
plotBFDirectionWetzels(BFWetzels, elicit.stage, disttype)
}
\arguments{
\item{BFWetzels}{Dataset containing Bayes factors for Wetzels data}

\item{elicit.stage}{Either "MATCH" or "Shiny"}

\item{disttype}{Either "t" or "norm"}
}
\description{
Function to plot the agreement in direction between Bayes factors for the Wetzels data
}
