% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BF.R
\name{plotBFWetzelsES}
\alias{plotBFWetzelsES}
\title{Plot Bayes factors for increasing sample effect sizes in the Wetzels et al data}
\usage{
plotBFWetzelsES(
  WetzelsEtAl,
  BFWetzels,
  ExpertsPriors_tTest,
  elicit.stage,
  alternative,
  disttype,
  percent = 0.3
)
}
\arguments{
\item{WetzelsEtAl}{Wetzels et al. dataset}

\item{BFWetzels}{Dataset containing Bayes factors for Wetzels data}

\item{ExpertsPriors_tTest}{Dataset ExpertsPriors_tTest containing prior distribution data for the t-test}

\item{elicit.stage}{"Shiny" or "MATCH" for elicited priors from Shiny app or MATCH}

\item{alternative}{one out of "two.sided", "greater", or "less"}

\item{disttype}{Which prior distribution family? "norm" for normal distribution, "t" for t-distribution}

\item{percent}{Determines xlim bounds. Credible interval of the prior distributions.}
}
\description{
Plot Bayes factors for increasing sample effect sizes in the Wetzels et al data
}
