% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BFAgreement.R
\name{plotEvidenceChangeBosco}
\alias{plotEvidenceChangeBosco}
\title{Function to plot the agreement in evidence strength between Bayes factors for the Bosco data}
\usage{
plotEvidenceChangeBosco(BFBosco, elicit.stage, threshold)
}
\arguments{
\item{BFBosco}{Dataset containing Bayes factors for Bosco data}

\item{elicit.stage}{Either "MATCH" or "Shiny"}

\item{threshold}{BF threshold}
}
\description{
Function to plot the agreement in evidence strength between Bayes factors for the Bosco data
}
