% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_KL.R
\name{plotKLBoscoES}
\alias{plotKLBoscoES}
\title{Plot KL for Bosco et al. data}
\usage{
plotKLBoscoES(
  BoscoEtAl,
  KLBosco,
  ExpertsPriors_cor,
  elicit.stage,
  percent = 0.3
)
}
\arguments{
\item{BoscoEtAl}{Wetzels et al. dataset}

\item{KLBosco}{Dataset containing Bayes factors for Wetzels data}

\item{ExpertsPriors_cor}{Dataset ExpertsPriors_tTest containing prior distribution data for the t-test}

\item{elicit.stage}{"Shiny" or "MATCH" for elicited priors from Shiny app or MATCH}

\item{percent}{Determines xlim bounds. Credible interval of the prior distributions.}
}
\description{
Plot KL for Bosco et al. data
}
