% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_plot_KL.R
\name{plotKLWetzels}
\alias{plotKLWetzels}
\title{Plot BF for the Wetzels et al dataset}
\usage{
plotKLWetzels(
  KLWetzels,
  disttype = c("norm", "t"),
  elicit.stage = c("MATCH", "Shiny"),
  pdf.width = 8.5,
  pdf.height = 4,
  cex.text = 1.5,
  mar = c(5, 3, 1, 1),
  line.ylab = 1.5,
  xtck = NULL
)
}
\arguments{
\item{KLWetzels}{Dataset containing BF of Wetzels et al. data}

\item{disttype}{Which type of prior distribution? "norm" = normal, "t" = t-distribution}

\item{elicit.stage}{Which stage of the elicitation process? "MATCH" = MATCH results, "Shiny" = Shiny results}

\item{pdf.width}{Width of the pdf output}

\item{pdf.height}{Height of the pdf output}

\item{cex.text}{Font size of axis labels}

\item{mar}{Margins of the plot to be inserted in par(mar=...)}

\item{line.ylab}{Position of y-axis label relative to y axis (to be inserted in mtext(line=...))}

\item{xtck}{Vector specifying location of tick marks on x-axis, automatically determined if xtcks=NULL (default)}
}
\description{
Plot BF for the Wetzels et al dataset
}
