% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_KL_ES.R
\name{plotKLWetzelsES}
\alias{plotKLWetzelsES}
\title{Plot KL divergence for Wetzels et al. data}
\usage{
plotKLWetzelsES(
  WetzelsEtAl,
  KLWetzels,
  ExpertsPriors_tTest,
  elicit.stage,
  disttype,
  percent = 0.3
)
}
\arguments{
\item{WetzelsEtAl}{Wetzels et al. dataset}

\item{KLWetzels}{Dataset containing Bayes factors for Wetzels data}

\item{ExpertsPriors_tTest}{Dataset ExpertsPriors_tTest containing prior distribution data for the t-test}

\item{elicit.stage}{"Shiny" or "MATCH" for elicited priors from Shiny app or MATCH}

\item{disttype}{Which prior distribution family? "norm" for normal distribution, "t" for t-distribution}

\item{percent}{Determines xlim bounds. Credible interval of the prior distributions.}
}
\description{
Plot KL divergence for Wetzels et al. data
}
