% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_priors.R
\name{plotpriors_cor}
\alias{plotpriors_cor}
\title{Plot Prior Distributions for t-test}
\usage{
plotpriors_cor(ExpertsPriors_cor, ylim)
}
\arguments{
\item{ExpertsPriors_cor}{Expert prior dataset for correlation (see /data-raw)}

\item{ylim}{Vector of length 2; limits of the y axis}
}
\description{
Plot Prior Distributions for t-test
}
