% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_priors.R
\name{plotpriors_t}
\alias{plotpriors_t}
\title{Plot Prior Distributions for t-test}
\usage{
plotpriors_t(ExpertsPriors_tTest, ylim, disttype)
}
\arguments{
\item{ExpertsPriors_tTest}{Expert prior dataset for t-test (see /data-raw)}

\item{ylim}{Vector of length 2; limits of the y axis}

\item{disttype}{Normal distribution ("norm") or t-distribution ("t")}
}
\description{
Plot Prior Distributions for t-test
}
