% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BFAgreement.R
\name{plotEvidenceChangeWetzels}
\alias{plotEvidenceChangeWetzels}
\title{Function to plot the agreement in evidence strength between Bayes factors for the Wetzels data}
\usage{
plotEvidenceChangeWetzels(BFWetzels, elicit.stage, threshold, disttype)
}
\arguments{
\item{BFWetzels}{Dataset containing Bayes factors for Bosco data}

\item{elicit.stage}{Either "MATCH" or "Shiny"}

\item{threshold}{BF threshold}

\item{disttype}{Either "t" or "norm"}
}
\description{
Function to plot the agreement in evidence strength between Bayes factors for the Wetzels data
}
